Add light direction and mainloop()
The direction of light can now be set in the same way as the viewAxis,
though the distance to the camera and the light source are the same.
ALso, all lines are now trimmed down to 79 characters in length and
mainloop is added to the end to accomodate non-IDLE editors.