Put all colour settings in a big list
The colours of vertices, lines, faces, the sphere, and menus
can now be set by modifying the COLOURS constant.
Colours in the polytope are represented by integers that assign
a number to each object; each number is later mapped to a colour.
Everything that controls light, and colour, are in single functions.