Add light colour and camera position scales
There are now three sections of scales which scale most scalable stuff,
including light properties, light colours, and camera position angles.
There are no scales for the rotation axis-plane because the concept is
unintuitive as a scale, and the user can just move the camera instead.
Since lighting is only allowed in only3D mode, lomega is not necessary.
Sliders have to communicate with the canvas very stupidly at the moment
and it would be nice to reorganize those properties to put them in main.
Also, I realized while sliding the camera that my view algorithm sucks.